server/requirements.txt
fastapi
uvicorn[standard]
gunicorn
sqlalchemy
psycopg2-binary
passlib[bcrypt]
PyJWT
celery
kombu[sqlalchemy]
ansible-runner
paramiko
python-dotenv
